+++ configs/recommender.yaml
@@ -1,0 +1,16 @@
+name: "ml-20m 2000 Users"
+random_seed: 42
+batch_size: 1024
+
+recommender:
+  embed_dim: 768
+  mlp_dims: [768, 768]
+  dropout: 0.8
+  weights: NULL
+
+optimizer:
+  lr: 0.001
+  weight_decay: 0.01
+  
+train:
+  max_epochs: 10
